♻ Refactor: 댓글 및 답글 작성 Component #31
AuthorInfo.tsx 파일 위치 폴더명 수정
불필요한 주석 삭제
