+++ src/main/reactfront/src/components/common/authorInfo/AuthorInfo.tsx
